+++ StMoviePlayer/StMoviePlayerGUI.cpp
@@ -1892,7 +1892,7 @@ void StMoviePlayerGUI::doListHotKeys(const size_t ) {
     StGLTable* aTable = new StGLTable(aDialog->getContent(), 0, 0, StGLCorner(ST_VCORNER_TOP, ST_HCORNER_CENTER));
     aTable->changeItemMargins().top    = scale(4);
     aTable->changeItemMargins().bottom = scale(4);
-    aTable->setupTable(anActionsMap.size(), 3);
+    aTable->setupTable((int )anActionsMap.size(), 3);
 
     StHandle< StSlot<void (const size_t )> > aSlot1 = new StSlotMethod<StMoviePlayerGUI, void (const size_t )>(this, &StMoviePlayerGUI::doChangeHotKey1);
     StHandle< StSlot<void (const size_t )> > aSlot2 = new StSlotMethod<StMoviePlayerGUI, void (const size_t )>(this, &StMoviePlayerGUI::doChangeHotKey2);
@@ -1905,13 +1905,13 @@ void StMoviePlayerGUI::doListHotKeys(const size_t ) {
 }
 
 void StMoviePlayerGUI::doChangeHotKey1(const size_t theId) {
-    const StHandle<StAction>& anAction = myPlugin->getAction(theId);
+    const StHandle<StAction>& anAction = myPlugin->getAction((int )theId);
     StHotKeyControl* aKeyChanger = new StHotKeyControl(myPlugin, myHKeysTable, this, anAction, 1);
     aKeyChanger->stglInit();
 }
 
 void StMoviePlayerGUI::doChangeHotKey2(const size_t theId) {
-    const StHandle<StAction>& anAction = myPlugin->getAction(theId);
+    const StHandle<StAction>& anAction = myPlugin->getAction((int )theId);
     StHotKeyControl* aKeyChanger = new StHotKeyControl(myPlugin, myHKeysTable, this, anAction, 2);
     aKeyChanger->stglInit();
 }
